Added new Volunteer Component with App Bar to Approvals, Applications, Progress
